# specs/002-enterprise-uiux-overhaul/contracts/gateway-uiux.openapi.yaml
openapi: 3.0.3
info:
  title: Orbit Tasks Gateway v2 Contract
  version: 2.0.0
  description: Public REST contract for Sprint-DSU loop, work item multiview CRUD, collaboration triage, and schedule intelligence.
servers:
  - url: https://tasksapi.infinitefallcult.trade
    description: Production gateway
  - url: http://localhost:8080
    description: Local gateway
paths:
...
  /api/v2/teams:
    post:
      summary: Create team
      operationId: createTeamV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, name, createdBy]
              properties:
                workspaceId:
                  type: string
                  format: uuid
                name:
                  type: string
                createdBy:
                  type: string
      responses:
        "200":
          description: Team
  /api/v2/teams/{teamId}/members:
    get:
      summary: List team members
      operationId: listTeamMembersV2
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team members
  /api/v2/teams/{teamId}/invites:
    get:
      summary: List team invites
      operationId: listTeamInvitesV2
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team invites
    post:
      summary: Create team invite
      operationId: inviteTeamMemberV2
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invitee, role, invitedBy]
              properties:
                invitee:
                  type: string
                role:
                  type: string
                invitedBy:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Invite
  /api/v2/teams/{teamId}/members/{userId}:
    patch:
      summary: Patch team member role
      operationId: patchTeamMemberV2
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, actorRole]
              properties:
                role:
                  type: string
                actorRole:
                  type: string
                note:
                  type: string
      responses:
        "200":
          description: Member
  /api/v2/insights/evaluations:
    post:
      summary: Run schedule evaluation
      operationId: runScheduleEvaluationV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleEvaluationRequest"
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEvaluation"
  /api/v2/insights/evaluations/{evaluationId}/actions:
    post:
      summary: Submit evaluation action
      operationId: submitEvaluationActionV2
      parameters:
        - in: path
          name: evaluationId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                note:
                  type: string
                patch:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Action result
components:
...
  schemas:
...
    ScheduleEvaluationRequest:
      type: object
      required: [workspaceId, projectId, remainingStoryPoints, availableCapacitySp, blockedCount, atRiskCount]
      properties:
        workspaceId:
          type: string
        projectId:
          type: string
        sprintId:
          type: string
        selectedWorkItemId:
          type: string
        prompt:
          type: string
        remainingStoryPoints:
          type: integer
        availableCapacitySp:
          type: integer
        blockedCount:
          type: integer
        atRiskCount:
          type: integer
        simulateAiFailure:
          type: boolean
    ScheduleEvaluation:
      type: object
      properties:
        evaluationId:
          type: string
        health:
          type: string
        topRisks:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              summary:
                type: string
              impact:
                type: string
              recommendedActions:
                type: array
                items:
                  type: string
              evidence:
                type: array
                items:
                  type: string
        questions:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: object
            properties:
              actionId:
                type: string
              label:
                type: string
              status:
                type: string
              note:
                type: string
        confidence:
          type: number
        fallback:
          type: boolean
        reason:
          type: string